Using HTTPoison.Base, and cleaning up some uneeded changes
1) HTTPoison.Base already defines get, post, put, delete and calls
request. I added a use HTTPoison.Base and overrided the request
method, and at the end called it directly. Since we need the
method in the headers definition there is no way around this.

2) Avoid using Dict.* if possible. I changed it to just use
Map.merge, in this case it makes a huge difference because if
headers where a keyword list the merge would be O(n^2).
Luckily for us you expect a map from the caller and we can assume
you will be the caller so map it is!

3) HTTPoison will also convert Headers from a map to a list see
https://github.com/edgurgel/httpoison/blob/master/lib/httpoison/base.ex#L91
so we can skip that step.

I don't have the xero key so I couldn't test this locally but it
builds!